Frd 267 interfaces payments admin

+++ backend/src/main/java/com/desertakal/desertakal/model/dto/admin/AdminDashboardDTO.java
@@ -3,6 +3,7 @@ package com.desertakal.desertakal.model.dto.admin;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -12,7 +13,7 @@ public class AdminDashboardDTO {
     private long totalUsers;
     private long totalTours;
     private long totalReservations;
-    private double totalRevenue;
+    private BigDecimal totalRevenue;
     private double averageTourRating;
 
     private long totalArticles;
